requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
# Optional RAG extras (uncomment to use)
langchain-cohere
langchain-mongodb
pymongo
